+++ broker/broker_utils/broker_utils/__init__.py
@@ -4,4 +4,4 @@
 """Tools to used by the Pitt-Google Broker software.
 """
 
-from . import consumer_sim, data_utils, gcp_utils, math, schema_maps, testing, types
+from . import avro_schemas, consumer_sim, data_utils, gcp_utils, math, schema_maps, testing, types